+++ backend/requirements.txt
@@ -1,0 +1,4 @@
+Django~=4.2.11
+Faker~=21.0.1
+djangorestframework~=3.14.0
+django-filter~=23.5
