--- requirements.txt
@@ -1,241 +1,0 @@
-alembic==1.13.1
-appdirs==1.4.4
-argcomplete==2.0.0
-asgiref==3.8.1
-astroid==3.0.3
-Authlib==1.2.1
-azure-common==1.1.28
-azure-core==1.28.0
-azure-identity==1.10.0
-azure-mgmt-core==1.3.2
-azure-mgmt-rdbms==10.2.0b12
-azure-mgmt-resource==23.1.0b2
-azure-mgmt-subscription==3.0.0
-Babel==2.12.1
-bcrypt==4.0.1
-Beaker==1.12.1
-beautifulsoup4==4.12.3
-bidict==0.22.1
-binwalk==2.3.4
-blinker==1.7.0
-blivet==3.8.2
-blivet-gui==2.5.0
-boto3==1.34.47
-botocore==1.34.47
-Brlapi==0.8.5
-Brotli==1.1.0
-cachetools==5.3.1
-capstone==4.0.2
-certifi==2023.5.7
-cffi==1.15.1
-chardet==5.2.0
-charset-normalizer==3.2.0
-click==8.1.3
-colored-traceback==0.3.0
-contourpy==1.2.0
-cryptography==41.0.7
-cson==0.8
-cssselect==1.1.0
-cupshelpers==1.0
-cycler==0.11.0
-dasbus==1.7
-dbus-python==1.3.2
-decorator==5.1.1
-defusedxml==0.7.1
-dill==0.3.8
-distlib==0.3.6
-distro==1.8.0
-Django==4.2.11
-django-cors-headers==4.3.1
-django-crispy-forms==2.1
-django-filter==23.5
-django-templated-mail==1.1.1
-djangorestframework==3.14.0
-djangorestframework-simplejwt==5.3.1
-djoser==2.2.2
-dnf==4.19.2
-dnspython==2.4.2
-email-validator==2.1.1
-eventlet==0.33.3
-Faker==21.0.1
-fb-re2==1.0.7
-fedora-third-party==0.10
-file-magic==0.4.0
-filelock==3.12.2
-Flask==2.2.5
-flask-babel==3.1.0
-Flask-Compress==1.14
-Flask-Gravatar==0.5.0
-Flask-Login==0.6.2
-Flask-Mail==0.9.1
-Flask-Migrate==4.0.7
-Flask-Paranoid==0.3.0
-Flask-Principal==0.4.0
-Flask-Security-Too==5.1.2
-Flask-SocketIO==5.3.6
-Flask-SQLAlchemy==3.0.5
-Flask-WTF==1.1.1
-fonttools==4.43.1
-fros==1.1
-fs==2.4.16
-google-api-core==2.11.1
-google-api-python-client==2.120.0
-google-auth==2.28.2
-google-auth-httplib2==0.2.0
-google-auth-oauthlib==0.8.0
-googleapis-common-protos==1.63.0
-greenlet==3.0.1
-grpcio==1.48.4
-grpcio-status==1.48.4
-gssapi==1.7.3
-h11==0.14.0
-hashID==3.1.4
-httpagentparser==1.9.5
-httplib2==0.21.0
-humanize==3.13.1
-idna==3.4
-inkex==1.3.0
-intervaltree==3.1.0
-isodate==0.6.1
-isort==5.13.2
-itsdangerous==2.1.2
-jaraco.classes==3.3.0
-jeepney==0.8.0
-Jinja2==3.1.3
-jmespath==1.0.1
-keyring==23.11.0
-kiwisolver==1.4.5
-koji==1.34.0
-langtable==0.0.65
-lazy-object-proxy==1.9.0
-ldap3==2.9.1
-libcomps==0.1.20
-libdnf==0.73.1
-lxml==4.9.3
-Mako==1.2.3
-MarkupSafe==2.1.3
-matplotlib==3.8.4
-mccabe==0.7.0
-mercurial==6.5.3
-meson==1.3.2
-more-itertools==8.13.0
-msal==1.24.1
-msal-extensions==1.0.0
-msrest==0.7.1
-munkres==1.1.2
-nftables==0.1
-numpy==1.24.4
-oauthlib==3.2.2
-olefile==0.46
-packaging==23.1
-paramiko==3.4.0
-passlib==1.7.4
-Paste==3.5.3
-pexpect==4.8.0
-pid==2.2.3
-pillow==10.3.0
-pipenv==2023.2.18
-platformdirs==3.9.1
-plumbum==1.7.2
-ply==3.11
-pooch==1.5.2
-portalocker==2.8.2
-productmd==1.38
-progressbar2==3.53.2
-protobuf==3.19.6
-psutil==5.9.5
-psycopg==3.1.10
-psycopg2==2.9.6
-ptyprocess==0.7.0
-pwntools==4.11.0
-pwquality==1.4.5
-pyasn1==0.5.1
-pyasn1-modules==0.3.0
-pycairo==1.25.1
-pycparser==2.20
-pycrypto==2.6.1
-pycups==2.0.1
-pycurl==7.45.2
-pyelftools==0.31
-pyenchant==3.2.2
-pygit2==1.14.0
-Pygments==2.15.1
-PyGObject==3.46.0
-PyJWT==2.8.0
-pykickstart==3.48
-pylint==3.0.4
-PyNaCl==1.5.0
-pyOpenSSL==23.2.0
-pyotp==2.9.0
-pyparsing==3.0.9
-pyparted==3.13.0
-PyQt5==5.15.10
-PyQt5-sip==12.13.0
-pyserial==3.5
-PySocks==1.7.1
-python-augeas==1.1.0
-python-dateutil==2.8.2
-python-engineio==4.9.0
-python-meh==0.51
-python-socketio==5.9.0
-python-utils==3.3.3
-python3-openid==3.2.0
-pytz==2024.1
-pyudev==0.24.1
-pyxdg==0.27
-PyYAML==6.0.1
-qrcode==7.4.2
-regex==2023.12.25
-requests==2.28.2
-requests-file==1.5.1
-requests-ftp==0.3.1
-requests-gssapi==1.2.3
-requests-oauthlib==1.3.1
-ROPGadget==7.3
-rpm==4.19.1.1
-rpmautospec==0.6.3
-rpmautospec-core==0.1.4
-rpyc==5.1.0
-rsa==4.9
-s3transfer==0.10.1
-scapy==2.5.0
-SciPy==1.11.1
-scour==0.38.2
-SecretStorage==3.3.3
-selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
-sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
-setools==4.4.3
-setuptools==67.7.2
-simple-websocket==1.0.0
-simpleaudio==1.0.4
-simplejson==3.19.2
-simpleline==1.9.0
-six==1.16.0
-social-auth-app-django==5.4.0
-social-auth-core==4.5.3
-sortedcontainers==2.4.0
-sos==4.6.0
-soupsieve==2.5
-speg==0.3
-SQLAlchemy==1.4.52
-sqlparse==0.4.2
-sshtunnel==0.4.0
-systemd-python==235
-Tempita==0.5.2
-tomli==2.0.1
-tomlkit==0.11.4
-typed-ast==1.5.5
-typing_extensions==4.8.0
-ua-parser==0.18.0
-unicorn==2.0.1.post1
-uritemplate==4.1.1
-urllib3==1.26.18
-user-agents==2.2.0
-virtualenv==20.21.1
-virtualenv-clone==0.5.7
-Werkzeug==2.2.3
-wfuzz==3.1.0
-wrapt==1.15.0
-wsproto==1.2.0
-WTForms==3.0.1
-zombie-imp==0.0.2
